exercise1_replication/requirements.txt
# sudata_de_challenge/requirements.txt
psycopg2-binary
pandas
python-dotenv
sqlalchemy